Feat: Final adaptations to template files and README

+++ init.sh
@@ -7,23 +7,16 @@ for templateFile in backend-config/*.template.*; do
 
   file=${templateFile/.template}	
 
-  if [ ! -f file ]; then 
-    cp ${templateFile} ${file}
-  fi
+  if [ ! -f file ]; then cp ${templateFile} ${file}; fi
 done	
 
 for templateFile in certs/*.template.*; do
 
   file=${templateFile/.template}	
 
-  if [ ! -f file ]; then 
-    cp ${templateFile} ${file}
-  fi
+  if [ ! -f file ]; then cp ${templateFile} ${file}; fi
 done	
 
 
-if [ ! -f .env ]; then 
-  cp .env.template .env
-fi
-
+if [ ! -f .env ]; then cp .env.template .env; fi
